moved the grisms reference file code to separate files
