+++ jwreftools/nircam/__init__.py
@@ -1,1 +1,2 @@
 from nircam_reftools import *
+from nircam_grism_reffiles import *
