+++ jwreftools/niriss/__init__.py
@@ -1,1 +1,2 @@
 from niriss_reftools import *
+from niriss_grism_reffiles import *
